+++ src/test/SuperFormFactory.t.sol
@@ -22,13 +22,9 @@ contract SuperFormFactoryTest is BaseSetup {
     /// @param formId is the id of the form
     /// @param vault is the address of the vault
     /// @param superFormId is the id of the superform - pair (form,vault)
-    event SuperFormCreated(
-        uint256 indexed formId,
-        address indexed vault,
-        uint256 indexed superFormId
-    );
+    event SuperFormCreated(uint256 indexed formId, address indexed vault, uint256 indexed superFormId);
 
-    uint16 internal chainId = ETH;
+    uint64 internal chainId = ETH;
 
     function setUp() public override {
         super.setUp();
@@ -37,38 +33,31 @@ contract SuperFormFactoryTest is BaseSetup {
     function test_chainId() public {
         vm.selectFork(FORKS[chainId]);
 
-        assertEq(
-            chainId,
-            ISuperRegistry(getContract(chainId, "SuperRegistry")).chainId()
-        );
+        assertEq(chainId, ISuperRegistry(getContract(chainId, "SuperRegistry")).chainId());
     }
 
     function test_revert_addForm_addressZero() public {
         address form = address(0);
-        uint256 formId = 1;
+        uint32 formId = 1;
 
         vm.prank(deployer);
         vm.expectRevert(Error.ZERO_ADDRESS.selector);
-        SuperFormFactory(getContract(chainId, "SuperFormFactory"))
-            .addFormBeacon(form, formId, salt);
+        SuperFormFactory(getContract(chainId, "SuperFormFactory")).addFormBeacon(form, formId, salt);
     }
 
     function test_revert_addForm_interfaceUnsupported() public {
         address form = address(0x1);
-        uint256 formId = 1;
+        uint32 formId = 1;
 
         vm.prank(deployer);
         vm.expectRevert(Error.ERC165_UNSUPPORTED.selector);
-        SuperFormFactory(getContract(chainId, "SuperFormFactory"))
-            .addFormBeacon(form, formId, salt);
+        SuperFormFactory(getContract(chainId, "SuperFormFactory")).addFormBeacon(form, formId, salt);
     }
 
     function test_addForm() public {
         vm.startPrank(deployer);
-        address formImplementation = address(
-            new ERC4626Form(getContract(chainId, "SuperRegistry"))
-        );
-        uint256 formBeaconId = 1;
+        address formImplementation = address(new ERC4626Form(getContract(chainId, "SuperRegistry")));
+        uint32 formBeaconId = 1;
 
         /*
         /// @dev FIXME: cannot predict address of beacon
@@ -81,8 +70,11 @@ contract SuperFormFactoryTest is BaseSetup {
         );
         emit FormCreated(form, 1);
         */
-        SuperFormFactory(getContract(chainId, "SuperFormFactory"))
-            .addFormBeacon(formImplementation, formBeaconId, salt);
+        SuperFormFactory(getContract(chainId, "SuperFormFactory")).addFormBeacon(
+            formImplementation,
+            formBeaconId,
+            salt
+        );
 
         //assertEq(formId, 1);
     }
@@ -90,8 +82,8 @@ contract SuperFormFactoryTest is BaseSetup {
     struct TestArgs {
         address formImplementation1;
         address formImplementation2;
-        uint256 formBeaconId1;
-        uint256 formBeaconId2;
+        uint32 formBeaconId1;
+        uint32 formBeaconId2;
         address vault1;
         address vault2;
         uint256 expectedSuperFormId1;
@@ -100,13 +92,13 @@ contract SuperFormFactoryTest is BaseSetup {
         address superForm;
         address resSuperForm;
         uint256 resFormid;
-        uint16 resChainId;
+        uint64 resChainId;
         uint256[] superFormIds_;
-        uint256[] formIds_;
-        uint16[] chainIds_;
+        uint32[] formIds_;
+        uint64[] chainIds_;
         uint256[] transformedChainIds_;
         uint256[] expectedSuperFormIds;
-        uint256[] expectedFormBeaconIds;
+        uint32[] expectedFormBeaconIds;
         uint256[] expectedChainIds;
         address[] superForms_;
         address[] expectedVaults;
@@ -121,38 +113,30 @@ contract SuperFormFactoryTest is BaseSetup {
             vm.selectFork(FORKS[chainIds[i]]);
             address superRegistry = getContract(chainId, "SuperRegistry");
             vars.formImplementation1 = address(new ERC4626Form(superRegistry));
-            vars.formImplementation2 = address(
-                new ERC4626TimelockForm(superRegistry)
-            );
+            vars.formImplementation2 = address(new ERC4626TimelockForm(superRegistry));
 
             vars.formBeaconId1 = 1;
             vars.formBeaconId2 = 2;
 
-            SuperFormFactory(getContract(chainId, "SuperFormFactory"))
-                .addFormBeacon(
-                    vars.formImplementation1,
-                    vars.formBeaconId1,
-                    salt
-                );
-            SuperFormFactory(getContract(chainId, "SuperFormFactory"))
-                .addFormBeacon(
-                    vars.formImplementation2,
-                    vars.formBeaconId2,
-                    salt
-                );
+            SuperFormFactory(getContract(chainId, "SuperFormFactory")).addFormBeacon(
+                vars.formImplementation1,
+                vars.formBeaconId1,
+                salt
+            );
+            SuperFormFactory(getContract(chainId, "SuperFormFactory")).addFormBeacon(
+                vars.formImplementation2,
+                vars.formBeaconId2,
+                salt
+            );
 
             /// @dev as you can see we are not testing if the vaults are eoas or actual compliant contracts
             vars.vault1 = address(0x2);
             vars.vault2 = address(0x3);
 
             /// @dev test getAllSuperForms
-            (
-                vars.superFormIds_,
-                vars.superForms_,
-                vars.formIds_,
-                vars.chainIds_
-            ) = SuperFormFactory(getContract(chainId, "SuperFormFactory"))
-                .getAllSuperForms();
+            (vars.superFormIds_, vars.superForms_, vars.formIds_, vars.chainIds_) = SuperFormFactory(
+                getContract(chainId, "SuperFormFactory")
+            ).getAllSuperForms();
 
             vars.transformedChainIds_ = new uint256[](vars.chainIds_.length);
 
@@ -160,12 +144,8 @@ contract SuperFormFactoryTest is BaseSetup {
                 vars.transformedChainIds_[j] = uint256(vars.chainIds_[j]);
             }
 
-            vars.expectedFormBeaconIds = new uint256[](
-                chainIds.length * UNDERLYING_TOKENS.length
-            );
-            vars.expectedChainIds = new uint256[](
-                chainIds.length * UNDERLYING_TOKENS.length
-            );
+            vars.expectedFormBeaconIds = new uint32[](chainIds.length * UNDERLYING_TOKENS.length);
+            vars.expectedChainIds = new uint256[](chainIds.length * UNDERLYING_TOKENS.length);
 
             /*
         for (uint256 i = 0; i < vars.expectedFormBeaconIds.length; i++) {
@@ -179,13 +159,10 @@ contract SuperFormFactoryTest is BaseSetup {
 
             /// @dev 6 * 2 * 3 = 36
 
-            uint256 expectedNumberOfSuperforms = chainIds.length *
-                UNDERLYING_TOKENS.length *
-                FORM_BEACON_IDS.length;
+            uint256 expectedNumberOfSuperforms = chainIds.length * UNDERLYING_TOKENS.length * FORM_BEACON_IDS.length;
 
             assertEq(
-                SuperFormFactory(getContract(chainIds[i], "SuperFormFactory"))
-                    .getAllSuperFormsList(),
+                SuperFormFactory(getContract(chainIds[i], "SuperFormFactory")).getAllSuperFormsList(),
                 expectedNumberOfSuperforms
             ); /// 1 chain = 3; 6 chains = 6 * 3
         }
@@ -194,12 +171,11 @@ contract SuperFormFactoryTest is BaseSetup {
     function xtest_pauseFormBeacon() public {
         vm.startPrank(deployer);
         vm.selectFork(FORKS[chainId]);
-        uint256 formBeaconId = 1;
+        uint32 formBeaconId = 1;
 
         vm.recordLogs();
         /// setting the status as false in chain id = ETH & broadcasting it
-        SuperFormFactory(getContract(chainId, "SuperFormFactory"))
-            .changeFormBeaconPauseStatus{value: 800 * 10 ** 18}(
+        SuperFormFactory(getContract(chainId, "SuperFormFactory")).changeFormBeaconPauseStatus{value: 800 * 10 ** 18}(
             formBeaconId,
             true,
             generateBroadcastParams(5, 2)
@@ -211,17 +187,13 @@ contract SuperFormFactoryTest is BaseSetup {
             if (chainIds[i] != chainId) {
                 vm.selectFork(FORKS[chainIds[i]]);
 
-                bool statusBefore = SuperFormFactory(
-                    payable(getContract(chainIds[i], "SuperFormFactory"))
-                ).getFormBeaconStatus(1);
+                bool statusBefore = SuperFormFactory(payable(getContract(chainIds[i], "SuperFormFactory")))
+                    .getFormBeaconStatus(1);
 
-                FactoryStateRegistry(
-                    payable(getContract(chainIds[i], "FactoryStateRegistry"))
-                ).processPayload(31, "");
+                FactoryStateRegistry(payable(getContract(chainIds[i], "FactoryStateRegistry"))).processPayload(31, "");
 
-                bool statusAfter = SuperFormFactory(
-                    payable(getContract(chainIds[i], "SuperFormFactory"))
-                ).getFormBeaconStatus(1);
+                bool statusAfter = SuperFormFactory(payable(getContract(chainIds[i], "SuperFormFactory")))
+                    .getFormBeaconStatus(1);
 
                 /// assert status update before and after processing the payload
                 assertEq(statusBefore, false);
